Refactor an example routing to handle root and table pages
