+++ packages/pxweb2/src/main.tsx
@@ -1,23 +1,21 @@
 import { StrictMode, Suspense } from 'react';
 import * as ReactDOM from 'react-dom/client';
-import { createBrowserRouter, Navigate, RouterProvider } from 'react-router';
+import { createBrowserRouter, RouterProvider } from 'react-router';
 
 import './i18n/config';
-import App from './app/app';
+
 import { validateConfig } from './app/util/validate';
-import { VariablesProvider } from './app/context/VariablesProvider';
-import { TableDataProvider } from './app/context/TableDataProvider';
-import { AccessibilityProvider } from './app/context/AccessibilityProvider';
-import { AppProvider } from './app/context/AppProvider';
+import TableView from './app/TableView/TableView';
+import App from './app/app';
 
 const router = createBrowserRouter([
   {
     path: '/',
-    element: <Navigate to="/table/tab638" replace={true} />,
+    element: <App />,
   },
   {
     path: '/table/:tableId',
-    element: <App />,
+    element: <TableView />,
   },
 ]);
 
@@ -31,16 +29,8 @@ if (location.href.includes('localhost')) {
 
 root.render(
   <StrictMode>
-    <AppProvider>
-      <AccessibilityProvider>
-        <VariablesProvider>
-          <TableDataProvider>
-            <Suspense fallback={<div>Loading...</div>}>
-              <RouterProvider router={router} />
-            </Suspense>
-          </TableDataProvider>
-        </VariablesProvider>
-      </AccessibilityProvider>
-    </AppProvider>
+    <Suspense fallback={<div>Loading...</div>}>
+      <RouterProvider router={router} />
+    </Suspense>
   </StrictMode>,
 );
